Move opa to forwarder
