+++ processor/forwarder.go
@@ -3,8 +3,11 @@ package processor
 import (
 	"encoding/json"
 	"fmt"
+	"strings"
 
+	"github.com/Financial-Times/go-logger/v2"
 	"github.com/Financial-Times/kafka-client-go/v4"
+	"github.com/Financial-Times/post-publication-combiner/v2/policy"
 )
 
 const (
@@ -18,9 +21,11 @@ type messageProducer interface {
 type forwarder struct {
 	producer              messageProducer
 	supportedContentTypes []string
+	log                   *logger.UPPLogger
+	opaAgent              policy.Agent
 }
 
-func newForwarder(producer messageProducer, supportedContentTypes []string) *forwarder {
+func newForwarder(producer messageProducer, supportedContentTypes []string, log *logger.UPPLogger, opaAgent policy.Agent) *forwarder {
 	return &forwarder{
 		producer:              producer,
 		supportedContentTypes: supportedContentTypes,
@@ -28,12 +33,32 @@ func newForwarder(producer messageProducer, supportedContentTypes []string) *for
 }
 
 func (f *forwarder) filterAndForwardMsg(headers map[string]string, message *CombinedModel) error {
+
+	tid := message.UUID
+	log := f.log.
+		WithTransactionID(tid).
+		WithField("processor", "forwarder")
+
 	if message.Content != nil {
 		contentType := message.Content.getType()
 
 		if !f.isTypeAllowed(contentType) {
 			return fmt.Errorf("%w: %s", ErrInvalidContentType, contentType)
 		}
+	}
+
+	result, err := f.opaAgent.EvaluateKafkaIngestPolicy(
+		message.Content,
+		policy.KafkaIngestMetadata,
+	)
+	if err != nil {
+		log.WithError(err).
+			Error("Could not evaluate the OPA Kafka Ingest policy while processing a /content/ metadata message.")
+		return err
+	}
+	if result.Skip {
+		log.Error(formatOPASkipReasons(result.Reasons))
+		return err
 	}
 
 	if err := f.forwardMsg(headers, message); err != nil {
@@ -63,4 +88,8 @@ func (f *forwarder) forwardMsg(headers map[string]string, message *CombinedModel
 		Headers: headers,
 		Body:    string(b),
 	})
+}
+
+func formatOPASkipReasons(r []string) string {
+	return strings.Join(r[:], ", ")
 }
